@@ -1,26 +1,42 @@
 cmake_minimum_required(VERSION 3.13)
 project(CppBasic2)
 
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 14)
 
 set(CMAKE_BUILD_TYPE Debug)
 
 #add_executable(CppBasic2 main.cpp)
 #add_executable(CppBasic_09_01 src/d09_template/dm01_func.cpp)
-#add_executable(CppBasic_09_02 src/d09_template/dm02_clazz.cpp)
-#add_executable(CppBasic_10_01 src/d10_cast/dm01_simple_cast.cpp)
+#add_executable(CppBasic_09_02 src/d09_template/dm02_fun_diff.cpp)
+#add_executable(CppBasic_09_03 src/d09_template/dm03_clazz.cpp)
+#add_executable(CppBasic_09_04 src/d09_template/dm04_clazz2.cpp)
+
+#add_executable(CppBasic_10_01 src/d10_cast/dm01_static_cast.cpp)
 #add_executable(CppBasic_10_02 src/d10_cast/dm02_dynamic_cast.cpp)
 #add_executable(CppBasic_10_03 src/d10_cast/dm03_const_cast.cpp)
+
 #add_executable(CppBasic_11_01 src/d11_exception/dm01_basic_exception.cpp)
 #add_executable(CppBasic_11_02 src/d11_exception/dm02_unwinding.cpp)
 #add_executable(CppBasic_11_03 src/d11_exception/dm03_lifecycle.cpp)
-#add_executable(CppBasic_11_04 src/d11_exception/dm04_test.cpp)
-#add_executable(CppBasic_12_01 src/d12_file/dm01_io.cpp)
-#add_executable(CppBasic_12_02 src/d12_file/dm02_binary_file.cpp)
+#add_executable(CppBasic_11_04 src/d11_exception/dm04_std_exception.cpp)
+#add_executable(CppBasic_12_01 src/d12_iostream/dm01_io.cpp)
+#add_executable(CppBasic_12_02 src/d12_iostream/dm02_binary_file.cpp)
 #add_executable(CppBasic_opus src/opus/ogg_pasrser.cpp src/opus/ogg_pasrser.h src/opus/oggs_page.cpp src/opus/oggs_page.h src/opus/ogg_app.cpp)
 #add_executable(CppBasic_13_01 src/d13_std/dm01_container.cpp)
 #add_executable(CppBasic_13_02 src/d13_std/dm02_string.cpp)
-#add_executable(CppBasic_14_01 src/d14_vector/dm01_vector.cpp)
+
+add_executable(CppBasic_14_01 src/d14_container/dm01_vector.cpp)
+#add_executable(CppBasic_14_02 src/d14_container/dm02_stack.cpp)
+#add_executable(CppBasic_14_03 src/d14_container/dm03_queue.cpp)
+#add_executable(CppBasic_14_04 src/d14_container/dm04_list.cpp)
+#add_executable(CppBasic_14_05 src/d14_container/dm05_priortity_queue.cpp)
+#add_executable(CppBasic_14_06 src/d14_container/dm06_set.cpp)
+#add_executable(CppBasic_14_07 src/d14_container/dm07_set_obj.cpp)
+#add_executable(CppBasic_14_08 src/d14_container/dm08_multi_set.cpp)
+#add_executable(CppBasic_14_09 src/d14_container/dm09_map.cpp)
+#add_executable(CppBasic_14_09 src/d14_container/dm10_multi_map.cpp)
+
+
 #add_executable(CppBasic_15_01 src/d15_set/dm01_set.cpp)
 #add_executable(CppBasic_15_02 src/d15_set/dm02_functor.cpp)
-add_executable(CppBasic_16_01 src/d16_map/dm01_basic.cpp)
+#add_executable(CppBasic_16_01 src/d16_map/dm01_basic.cpp)
