@@ -3,10 +3,11 @@
 //
 #include <iostream>
 #include <string>
+#include "algorithm"
 
 using namespace std;
 
-void test1() {
+void create_string() {
     string s1 = "aaaaaa";
     string s2("bbbbb");
     string s3 = s2; //通过拷贝构造函数 初始化s3；
@@ -19,7 +20,7 @@ void test1() {
 }
 
 //string 的迭代
-void test2() {
+void iterator_string() {
     //1.使用for循环迭代；
     string s = "abcdefg";
     for (int i = 0; i < s.size(); i++) {
@@ -28,43 +29,86 @@ void test2() {
     cout << endl;
 
     //2.使用迭代器
-    for (string::iterator it = s.begin(); it != s.end(); it++) {
+//    for (string::iterator it = s.begin(); it != s.end(); it++) {
+    for (auto it = s.begin(); it != s.end(); it++) {
         cout << *it << " ";
+    }
+    cout << endl;
+
+    //3.迭代器简化方法
+    for (char &it: s) {
+        cout << it << " ";
     }
     cout << endl;
 }
 
 //char * to string and string to char*
-void test3() {
+void convert_char_string() {
     string s = "abcdefg";
     printf("s=%s\n", s.c_str());
 
+    //string copy to buffer;
     char buf[128] = {0};
     s.copy(buf, 3, 0);
     printf("buf:%s\n", buf);
 }
 
 //字符串的拼接
-void test4() {
+void string_join() {
     string s1 = "aaaa";
     string s2 = "bbbb";
-    s1 += s2;
-    cout << "s1 = " << s1 << endl;
+    string s3 = s1 + s2;
+    cout << "s3 = " << s3 << endl;
     s1.append(s2);
     cout << "s1 = " << s1 << endl;
 }
 
 //find
-void test5() {
+void string_find() {
     string s = "hello world";
     int index = s.find("world", 0);
     cout << "index = " << index << endl;
 }
 
+void string_replace() {
+    string s = "hello world";
+    string replace = "world";
+    s.replace(s.find(replace), replace.length(), "WORLD");
+    cout << s << endl;
+}
+
+void sub_string() {
+    string s = "hello1 hello2 hello3";
+    auto it = find(s.begin(), s.end(), 'l');
+    if (it != s.end()) {
+        s.erase(it);
+    }
+
+    cout << s << endl;
+    s.erase(s.begin(), s.begin() + 5);
+    cout << s << endl;
+
+    s.insert(0, "new");
+    cout << s << endl;
+    s.insert(10, "new");
+    cout << s << endl;
+    //尾插法
+    s.insert(s.length(), "new");
+    cout << s << endl;
+}
+
+void string_normal() {
+    string s = "AAAA bbbb";
+//    std::transform(s.begin(), s.end(), s.begin(), std::tolower)
+}
+
 int main() {
-//    test1();
-//    test2();
-//    test3();
-//    test4();
-    test5();
+    create_string();
+    iterator_string();
+    convert_char_string();
+    string_join();
+    string_find();
+    string_replace();
+    sub_string();
+    string_normal();
 }
