@@ -4,9 +4,9 @@
 #include <iostream>
 
 template<typename T>
-class A {
+class Teacher {
 public:
-    A(T a) {
+    Teacher(T a) {
         this->a = a;
     }
 
@@ -22,9 +22,9 @@ private:
 /**
  * 子类派生模板父类必须制定类型；
  */
-class B : public A<int> {
+class MathTeacher : public Teacher<int> {
 public:
-    B(int a, int b) : A<int>(a) {
+    MathTeacher(int a, int b) : Teacher<int>(a) {
         this->b = b;
     }
 
@@ -37,19 +37,34 @@ private:
     int b;
 };
 
+template<typename T>
+class YuwenTeacher : public Teacher<T> {
+public:
+    YuwenTeacher(T c, T a) : Teacher<T>(a) {
+        this->c = c;
+    }
+
+    void printC(){
+        std::cout << "c:" << this->c << std::endl;
+    }
+private:
+    T c;
+};
+
 int main() {
     std::cout << "size of char = " << sizeof(char) << std::endl;
     std::cout << "size of short = " << sizeof(short) << std::endl;
     std::cout << "size of uint8_t = " << sizeof(uint8_t) << std::endl;
     std::cout << "size of char* = " << sizeof(char *) << std::endl;
-//    A<int> a(100);
-//    a.printA();
-//
-//    A<int> a2('c');
-//    a2.printA();
-//
-//
-//    B b(100, 200);
-//    b.printA();
-//    b.printB();
+
+    Teacher<int> a(100);
+    a.printA();
+
+    Teacher<int> a2('c');
+    a2.printA();
+
+
+    MathTeacher b(100, 200);
+    b.printA();
+    b.printB();
 }
