@@ -17,9 +17,14 @@ public:
         cout << "构造函数" << endl;
     }
 
+    CopyException(const CopyException &c) {
+        this->code = c.code;
+        this->msg = c.msg;
+        cout << "拷贝 构造函数" << endl;
+    }
 
     ~CopyException() {
-        cout << "析构函数" << endl;
+        cout << "~析构函数" << endl;
     }
 
     int getCode() {
@@ -28,12 +33,6 @@ public:
 
     string getMsg() {
         return this->msg;
-    }
-
-    CopyException(const CopyException &c) {
-        this->code = c.code;
-        this->msg = c.msg;
-        cout << "copy构造函数" << endl;
     }
 
 private:
@@ -60,11 +59,11 @@ int main() {
         char in[] = "abcdefg";
         char *out = nullptr;
         copy_switch(in, out);
+    } catch (CopyException *e) {
+        cout << "code *= " << e->getCode() << ",msg = " << e->getMsg() << endl;
     } catch (CopyException &e) {
         //e:就是copy构造函数。&e就少了一个copy的动作
         cout << "code &= " << e.getCode() << ",msg = " << e.getMsg() << endl;
-    } catch (CopyException *e) {
-        cout << "code *= " << e->getCode() << ",msg = " << e->getMsg() << endl;
     } catch (...) {
         cout << "unknown exception" << endl;
     }
